chore(dot/parachain): consistently use ParaID type

+++ dot/parachain/validation-protocol/statement_distribution_message_test.go
@@ -63,7 +63,7 @@ func TestStatementDistributionMessage(t *testing.T) {
 
 	secondedEnumValue := parachaintypes.Seconded{
 		Descriptor: parachaintypes.CandidateDescriptor{
-			ParaID:                      uint32(1),
+			ParaID:                      1,
 			RelayParent:                 hash5,
 			Collator:                    collatorID,
 			PersistedValidationDataHash: hash5,
